Delay in Breadcrumb Display after Clicking 'View Scenario (#1872)
Currently, when clicking 'View Scenario,' there is a delay in displaying the scenario name in the breadcrumb until the getScenario response is received. This delay shouldn’t occur, as we already know the scenario name upon selection and should be able to display it immediately in the breadcrumb. This ticket addresses the issue by allowing the breadcrumb to update without waiting for the API response.

Jira: https://sig-gis.atlassian.net/browse/PLAN-1776